test(table): separating expandable row test in coupel of test files for optimization
