+++ packages/core/src/components/table/table/test/expandable-row/interaction-1.e2e.ts-snapshots/tds-table-expandable-row-first-under-first-row-opened-expanded-row-with-text-Hello-world-1-1-linux.png
