+++ packages/core/src/components/table/table/test/expandable-row/interaction-2.e2e.ts-snapshots/tds-table-expandable-row-second-under-second-row-opened-expanded-row-with-text-Hello-to-you-too-1-linux.png
