+++ packages/core/src/components/table/table/test/expandable-row/interaction-3.e2e.ts-snapshots/tds-table-expandable-row-third-under-third-row-38f81--row-with-a-button-with-text-Call-to-action-1-linux.png
